now using u16 for texture handle

+++ lotus-script-sys/src/lib.rs
@@ -56,11 +56,11 @@ pub mod messages {
 pub mod textures {
     #[link(wasm_import_module = "textures")]
     extern "C" {
-        pub fn create(options: u64) -> u64;
-        pub fn add_action(texture: u64, options: u64);
-        pub fn get_pixel(texture: u64, x: u32, y: u32) -> u32;
-        pub fn apply_to(texture: u64, name: u64);
-        pub fn flush_actions(texture: u64) -> u32;
+        pub fn create(options: u64) -> u16;
+        pub fn add_action(texture: u16, options: u64);
+        pub fn get_pixel(texture: u16, x: u32, y: u32) -> u32;
+        pub fn apply_to(texture: u16, name: u64);
+        pub fn flush_actions(texture: u16) -> u32;
     }
 }
